pull out srpfm into sep file
